Move common utilities to utils.py
